feat: Support for conjugate gradient (CG) methods that support general matrix operations
The implementation of the conjugate gradient (CG) method has been updated to support any type of matrix operation, enhancing support for sparse matrices. The definitions of related structures and functions have been modified to accommodate the new `MatrixOp` feature, and test cases for sparse matrices have been added to ensure the correctness of the functionality.